requirements.txt
fastapi
uvicorn[standard]
pydantic
passlib[bcrypt]
python-multipart
argon2-cffi
PyJWT
geopy
APScheduler
python-dotenv
